+++ vuurmuur/main.h
@@ -381,7 +381,7 @@ int check_for_changed_dynamic_ips(struct vrmr_interfaces *interfaces);
 
 /* ruleset */
 int ruleset_add_rule_to_set(struct vrmr_list *, char *, char *,
-        unsigned long long, unsigned long long);
+        uint64_t, uint64_t);
 int load_ruleset(struct vrmr_ctx *);
 
 /* shape */
